Removed tests that test extractParams() functionality. We don't use it anymore
